Rename: viewmodel 폴더 생성
